+++ src/features/book-note/index.ts
@@ -1,1 +1,3 @@
 export { default as BookNoteForm } from './ui/BookNoteForm';
+export { default as BookNotePreview } from './ui/BookNotePreview';
+export { createBookNoteFormSchema, type BookNoteFormSchema } from './model/book-note-form-schema';
