+++ src/features/book-note/ui/BookNoteForm/index.tsx
@@ -1,17 +1,15 @@
 'use client';
 
-import { useCallback, useMemo } from 'react';
-import { useParams } from 'next/navigation';
+import { useCallback } from 'react';
 import { isNotNil } from 'es-toolkit';
 import { FieldErrors } from 'react-hook-form';
 import { DevTool } from '@hookform/devtools';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { usePersistentForm } from '@/shared/lib/form';
+import { UsePersistentFormReturn } from '@/shared/lib/form';
 import { FunnelProvider, useFunnel } from '@/shared/lib/funnel';
 import { Form } from '@/shared/ui/form';
 import { RenderCase } from '@/shared/ui/render-case';
-import { useBookDetail } from '@/entities/book';
-import { BookNoteFormSchema, createBookNoteFormSchema } from '../../model/book-note-form-schema';
+import { BookDetail } from '@/entities/book';
+import { BookNoteFormSchema } from '../../model/book-note-form-schema';
 import { stepFields } from '../../model/step-fields';
 import BookNoteFormActions from './BookNoteFormActions';
 import ReadingInfoStep from './ReadingInfoStep';
@@ -20,33 +18,12 @@ import ReviewStep from './ReviewStep';
 import QuotesStep from './QuotesStep';
 import PublishStep from './PublishStep';
 
-const defaultValues: BookNoteFormSchema = {
-  readingInfo: {
-    readingStatus: null,
-    startDate: null,
-    endDate: null,
-  },
-  recommended: null,
-  rating: null,
-  review: '',
-  quotes: [],
-  publish: false,
-};
+interface BookNoteFormProps {
+  book: BookDetail;
+  form: UsePersistentFormReturn<BookNoteFormSchema>;
+}
 
-export default function BookNoteForm() {
-  const { isbn } = useParams<{ isbn: string }>()!;
-
-  const book = useBookDetail(isbn);
-
-  const resolver = useMemo(() => zodResolver(createBookNoteFormSchema(book.pageCount)), [book.pageCount]);
-
-  const form = usePersistentForm<BookNoteFormSchema>({
-    resolver,
-    defaultValues,
-    storage: 'session',
-    storageKey: `book-note-form-${isbn}`,
-  });
-
+export default function BookNoteForm({ book, form }: BookNoteFormProps) {
   const onStepChange = useCallback(
     (step: number) => {
       for (const field of stepFields.get(step) ?? []) {
