+++ src/widgets/book-note/index.ts
@@ -1,0 +1,1 @@
+export { default as BookNoteEditor } from './ui/book-note-editor';
